// src/apprunner/models/apprunner.d.ts
    export interface AuthenticationConfiguration {
        ConnectionArn?: FusionResourceArn
        AccessRoleArn?: RoleArn
    }
// ...
    export type AutoScalingConfigurationName = string
    export type AutoScalingConfigurationStatus = 'ACTIVE' | 'INACTIVE' | string
    export interface AutoScalingConfigurationSummary {
        AutoScalingConfigurationArn?: FusionResourceArn
        AutoScalingConfigurationName?: AutoScalingConfigurationName
        AutoScalingConfigurationRevision?: Integer
    }
// ...
    export interface CodeRepository {
        RepositoryUrl: String
        SourceCodeVersion: SourceCodeVersion
        CodeConfiguration?: CodeConfiguration
    }
// ...
    export type Cpu = string
// ...
    export interface EncryptionConfiguration {
        KmsKey: KmsKeyArn
    }
    export type FusionResourceArn = string
    export interface HealthCheckConfiguration {
        Enabled?: NullableBoolean
        Path?: String
        Interval?: HealthCheckInterval
        Timeout?: HealthCheckTimeout
        HealthyThreshold?: HealthCheckHealthyThreshold
        UnhealthyThreshold?: HealthCheckUnhealthyThreshold
    }
    export type HealthCheckHealthyThreshold = number
    export type HealthCheckInterval = number
    export type HealthCheckTimeout = number
    export type HealthCheckUnhealthyThreshold = number
    export interface ImageConfiguration {
        RuntimeEnvironmentVariables?: RuntimeEnvironmentVariables
        StartCommand?: String
        Port?: String
    }
    export type ImageIdentifier = string
    export interface ImageRepository {
        ImageIdentifier: ImageIdentifier
        ImageConfiguration?: ImageConfiguration
        ImageRepositoryType: ImageRepositoryType
    }
    export type ImageRepositoryType = 'ECR' | 'ECR-Public' | string
    export type InitiatorType = 'USER' | 'SERVICE' | string
    export interface InstanceConfiguration {
        Cpu?: Cpu
        Memory?: Memory
        InstanceRoleArn?: RoleArn
    }
    export type Integer = number
    export type KmsKeyArn = string
// ...
    export type Memory = string
    export type NextToken = string
    export type NullableBoolean = boolean
    export type NullableInteger = number
    export interface Operation {
        Id?: UUID
        Type?: OperationType
        Status?: OperationStatus
        TargetArn?: FusionResourceArn
        StartedAt?: Timestamp
        UpdatedAt?: Timestamp
        EndedAt?: Timestamp
        InitiatorType?: InitiatorType
    }
    export type OperationStatus =
        | 'PENDING'
        | 'IN_PROGRESS'
        | 'FAILED'
        | 'SUCCEEDED'
        | 'ROLLBACK_IN_PROGRESS'
        | 'ROLLBACK_FAILED'
        | 'ROLLBACK_SUCCEEDED'
        | string
    export interface OperationSummary {
        Id?: UUID
        Type?: OperationType
        Status?: OperationStatus
        TargetArn?: FusionResourceArn
        StartedAt?: Timestamp
        EndedAt?: Timestamp
        UpdatedAt?: Timestamp
    }
    export type OperationSummaryList = OperationSummary[]
    export type OperationType = 'START_DEPLOYMENT' | 'CREATE_SERVICE' | 'PAUSE_SERVICE' | 'RESUME_SERVICE' | string
// ...
    export type RoleArn = string
    export type Runtime = 'python3' | 'nodejs10' | 'nodejs12' | string
    export type RuntimeEnvironmentVariables = { [key: string]: String }
    export interface Service {
        ServiceName: ServiceName
        ServiceId: ServiceId
        ServiceArn: FusionResourceArn
        ServiceUrl: String
        CreatedAt: Timestamp
        UpdatedAt: Timestamp
        DeletedAt?: Timestamp
        Status: ServiceStatus
        SourceConfiguration: SourceConfiguration
        InstanceConfiguration: InstanceConfiguration
        EncryptionConfiguration?: EncryptionConfiguration
        HealthCheckConfiguration?: HealthCheckConfiguration
        AutoScalingConfigurationSummary: AutoScalingConfigurationSummary
    }
    export type ServiceId = string
    export type ServiceMaxResults = number
    export type ServiceName = string
    export type ServiceStatus =
        | 'CREATE_FAILED'
        | 'RUNNING'
        | 'DELETED'
        | 'DELETE_FAILED'
        | 'PAUSED'
        | 'OPERATION_IN_PROGRESS'
        | string
    export interface ServiceSummary {
        ServiceName?: ServiceName
        ServiceId?: ServiceId
        ServiceArn?: FusionResourceArn
        ServiceUrl?: String
        CreatedAt?: Timestamp
        UpdatedAt?: Timestamp
        Status?: ServiceStatus
    }
    export type ServiceSummaryList = ServiceSummary[]
    export interface SourceCodeVersion {
        Type: SourceCodeVersionType
        Value: String
    }
    export type SourceCodeVersionType = 'Branch' | string
    export interface SourceConfiguration {
        CodeRepository?: CodeRepository
        ImageRepository?: ImageRepository
        AutoDeploymentsEnabled?: NullableBoolean
        AuthenticationConfiguration?: AuthenticationConfiguration
    }
    export interface StartDeploymentRequest {
        ServiceArn: FusionResourceArn
    }
    export interface StartDeploymentResponse {
        StartDeploymentResult: StartDeploymentResult
    }
    export interface StartDeploymentResult {
        OperationId: UUID
    }
    export type String = string
    export interface Tag {
        Key?: TagKey
        Value?: TagValue
    }
    export type TagKey = string
    export type TagKeyList = TagKey[]
    export type TagList = Tag[]
    export interface TagResourceRequest {
        ResourceArn: FusionResourceArn
        Tags: TagList
    }
    export interface TagResourceResponse {}
    export type TagValue = string
    export type Timestamp = Date
    export type UUID = string